+++ dev-dist/sw.js
@@ -82,7 +82,7 @@ define(['./workbox-21a80088'], (function (workbox) { 'use strict';
     "revision": "d3b122656a89942aa3164f5ccd56d9ca"
   }, {
     "url": "/Investment_calculator/offline.html",
-    "revision": "0.ufgh2sb2bv"
+    "revision": "0.6gdd1s8ia6g"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("/Investment_calculator/offline.html"), {